clean empty object extracted to utils
